rename host_functions module to crypto
