+++ tendermint/src/lib.rs
@@ -33,10 +33,10 @@ pub mod block;
 pub mod chain;
 pub mod channel;
 pub mod consensus;
+pub mod crypto;
 pub mod evidence;
 pub mod genesis;
 pub mod hash;
-pub mod host_functions;
 pub mod merkle;
 mod moniker;
 pub mod node;
